requirements.txt
discord.py==1.3.2
numpy==1.18.2
pandas==1.0.3
pre-commit==2.2.0
psycopg2-binary==2.8.4